# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/employee_taxes:
    get:
      tags:
        - EmployeeTaxes
      summary: List EmployeeTaxes
      description: Get a paginated list of EmployeeTaxes with optional filtering and sorting
      operationId: listEmployeeTaxes
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeTaxList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EmployeeTaxes
      summary: Create EmployeeTax
      description: Create a new EmployeeTax
      operationId: createEmployeeTax
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeTaxRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeTax'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/employee_work_experiences:
    get:
      tags:
        - EmployeeWorkExperiences
      summary: List EmployeeWorkExperiences
      description: Get a paginated list of EmployeeWorkExperiences with optional filtering and sorting
      operationId: listEmployeeWorkExperiences
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeWorkExperienceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EmployeeWorkExperiences
      summary: Create EmployeeWorkExperience
      description: Create a new EmployeeWorkExperience
      operationId: createEmployeeWorkExperience
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeWorkExperienceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeWorkExperience'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    EmployeeTax:
      type: object
      required:
        - id
        - employee_id
        - tax_method
        - tax_salary
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        npwp_number:
          nullable: true
          type: string
        ptkp_override:
          nullable: true
          type: string
        tax_method:
          type: string
        ter_category:
          nullable: true
          type: string
        tax_salary:
          type: string
        taxable_date:
          nullable: true
          type: string
          format: date
        beginning_netto:
          nullable: true
          type: string
        pph21_paid:
          nullable: true
          type: string
        metadata:
          type: object
...
    CreateEmployeeTaxRequest:
      type: object
      required:
        - employee_id
        - tax_method
        - tax_salary
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        npwp_number:
          nullable: true
          type: string
        ptkp_override:
          nullable: true
          type: string
        tax_method:
          type: string
        ter_category:
          nullable: true
          type: string
        tax_salary:
          type: string
        taxable_date:
          nullable: true
          type: string
          format: date
        beginning_netto:
          nullable: true
          type: string
        pph21_paid:
          nullable: true
          type: string
        metadata:
          type: object
...
    EmployeeTaxList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeTax'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    EmployeeWorkExperience:
      type: object
      required:
        - id
        - employee_id
        - company_name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        company_name:
          type: string
        job_position:
          nullable: true
          type: string
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    CreateEmployeeWorkExperienceRequest:
      type: object
      required:
        - employee_id
        - company_name
        - metadata
      properties:
        employee_id:
          type: string
          format: uuid
        company_name:
          type: string
        job_position:
          nullable: true
          type: string
        start_date:
          nullable: true
          type: string
          format: date
        end_date:
          nullable: true
          type: string
          format: date
        metadata:
          type: object
...
    EmployeeWorkExperienceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeWorkExperience'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []