# src/presentation/http/employee_tax_handler.rs
//! EmployeeTax REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{EmployeeTaxService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateEmployeeTaxDto, UpdateEmployeeTaxDto, PatchEmployeeTaxDto, EmployeeTaxResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum EmployeeTaxError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for EmployeeTaxError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for EmployeeTaxError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "EMPLOYEETAX_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "EMPLOYEETAX_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "EMPLOYEETAX_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "EMPLOYEETAX_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for EmployeeTax.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /employee_taxes | List with pagination |
/// | POST | /employee_taxes | Create new |
/// | GET | /employee_taxes/:id | Get by ID |
/// | PUT | /employee_taxes/:id | Full update |
/// | PATCH | /employee_taxes/:id | Partial update |
/// | DELETE | /employee_taxes/:id | Soft delete |
/// | POST | /employee_taxes/bulk | Bulk create |
/// | POST | /employee_taxes/upsert | Upsert |
/// | GET | /employee_taxes/trash | List deleted |
/// | POST | /employee_taxes/:id/restore | Restore |
/// | DELETE | /employee_taxes/empty | Empty trash |
/// | GET | /employee_taxes/:id/deleted | Get deleted by ID |
/// | DELETE | /employee_taxes/trash/:id | Permanent delete from trash |
/// | GET | /employee_taxes/count | Count active entities |
/// | GET | /employee_taxes/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(EmployeeTaxService::with_repository(repository));
/// let router = create_employee_tax_routes(service);
/// ```
pub fn create_employee_tax_routes(service: Arc<EmployeeTaxService>) -> Router {
    BackboneCrudHandler::<EmployeeTaxService, EmployeeTax, CreateEmployeeTaxDto, UpdateEmployeeTaxDto, EmployeeTaxResponseDto>::routes(
        service,
        "/employee_taxes",
    )
}

/// Create Axum router with only the read (GET) endpoints for EmployeeTax.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_employee_tax_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_employee_tax_read_routes(service: Arc<EmployeeTaxService>) -> Router {
    BackboneCrudHandler::<EmployeeTaxService, EmployeeTax, CreateEmployeeTaxDto, UpdateEmployeeTaxDto, EmployeeTaxResponseDto>::read_routes(
        service,
        "/employee_taxes",
    )
}

/// Create Axum router with only the write (mutation) endpoints for EmployeeTax.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_employee_tax_write_routes(service: Arc<EmployeeTaxService>) -> Router {
    BackboneCrudHandler::<EmployeeTaxService, EmployeeTax, CreateEmployeeTaxDto, UpdateEmployeeTaxDto, EmployeeTaxResponseDto>::write_routes(
        service,
        "/employee_taxes",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_employee_tax_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<EmployeeTaxService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_employee_tax_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}